fix: GTest::gtest_main, VkResult cast, GLFW_INCLUDE_VULKAN
Same set of fixes that landed on feat/gpu-swapchain via f3cd2e7 but
backported to this branch so PR #3 builds on its own.

- tests/CMakeLists.txt: link GTest::gtest_main (provides main() for the
  unit test executable)
- engine/src/gpu/{instance,physical_device}.cpp: cast VkResult to int
  before std::to_string — VkResult is an enum and std::to_string has
  no overload for it under strict C++23
- app/src/main.cpp: define GLFW_INCLUDE_VULKAN before <GLFW/glfw3.h>
  so glfwCreateWindowSurface is declared with Vulkan types

+++ engine/src/gpu/physical_device.cpp
@@ -57,7 +57,7 @@ auto PhysicalDevice::select(const Instance& instance, std::uint32_t min_api_vers
         vr != VK_SUCCESS) {
         return std::unexpected(noted::make_error(
             noted::ErrorCode::gpu_validation_failed,
-            std::string{"vkEnumeratePhysicalDevices(count) failed: "} + std::to_string(vr)));
+            std::string{"vkEnumeratePhysicalDevices(count) failed: "} + std::to_string(static_cast<int>(vr))));
     }
     if (count == 0) {
         return std::unexpected(noted::make_error(
@@ -70,7 +70,7 @@ auto PhysicalDevice::select(const Instance& instance, std::uint32_t min_api_vers
         vr != VK_SUCCESS) {
         return std::unexpected(noted::make_error(
             noted::ErrorCode::gpu_validation_failed,
-            std::string{"vkEnumeratePhysicalDevices failed: "} + std::to_string(vr)));
+            std::string{"vkEnumeratePhysicalDevices failed: "} + std::to_string(static_cast<int>(vr))));
     }
 
     PhysicalDevice best;
